Move the form into component
