+++ src/components/Todo.js
@@ -1,28 +1,16 @@
 import { useEffect, useState } from "react";
 import { API, graphqlOperation } from "aws-amplify";
 import { Container } from "@chakra-ui/react";
-import { createTodo } from "../graphql/mutations";
 import { listTodos } from "../graphql/queries";
-import SyntaxHighlighter from "react-syntax-highlighter";
-import { dark } from "react-syntax-highlighter/dist/esm/styles/hljs";
-import { VStack } from "@chakra-ui/react";
-import { Input } from "@chakra-ui/react";
-import { Textarea } from "@chakra-ui/react";
-import { Button, useColorModeValue } from "@chakra-ui/react";
-
-const initialState = { name: "", description: "" };
+import Card from "./Card";
+import PostForm from "./PostForm";
 
 export default function Todo() {
-  const [formState, setFormState] = useState(initialState);
   const [todos, setTodos] = useState([]);
-  
+
   useEffect(() => {
     fetchTodos();
   }, []);
-
-  function setInput(key, value) {
-    setFormState({ ...formState, [key]: value });
-  }
 
   async function fetchTodos() {
     try {
@@ -34,53 +22,13 @@ export default function Todo() {
     }
   }
 
-  async function addTodo() {
-    try {
-      if (!formState.name || !formState.description) return;
-      const todo = { ...formState };
-      setTodos([...todos, todo]);
-      setFormState(initialState);
-      await API.graphql(graphqlOperation(createTodo, { input: todo }));
-    } catch (err) {
-      console.log("error creating todo:", err);
-    }
-  }
   return (
     <Container maxW="container.md">
-      <VStack>
-        <h2 style={{paddingTop:'15px'}}>&lt;hoot your code/&gt;</h2>
-        <Input
-          style={{ width:"300px" }}
-          bg={useColorModeValue("")}
-          placeholder="Name"
-          value={formState.name}
-          onChange={(event) => setInput("name", event.target.value)}
-        />
-        <Textarea
-          style={{ height:"100px", overflowWrap: "break-word", overflow: "auto" }}
-          bg={useColorModeValue("")}
-          placeholder="Code"
-          value={formState.description}
-          onChange={(event) => setInput("description", event.target.value)}
-        />
-        <Button colorScheme="teal" onClick={addTodo}>
-          Hoot
-        </Button>
-      </VStack>
+      <PostForm newTodo={(newTodo) => setTodos([...todos, newTodo])}></PostForm>
 
       {todos.map((todo, index) => (
-        <div key={todo.id ? todo.id : index} style={styles.todo}>
-          <p style={styles.todoName}>{todo.name}</p>
-          <SyntaxHighlighter language="javascript" style={dark}>
-            {todo.description}
-          </SyntaxHighlighter>
-        </div>
+        <Card todo={todo} key={index}></Card>
       ))}
     </Container>
   );
 }
-
-const styles = {
-  todo: { marginBottom: 15 },
-  todoName: { fontSize: 20, fontWeight: "bold" },
-};
